add qrcode recognizing and server connection

+++ app/src/main/java/ru/shafran/cards/data/card/CardHistory.kt
@@ -1,12 +1,12 @@
 package ru.shafran.cards.data.card
 
-import com.arkivanov.decompose.statekeeper.Parcelable
-import com.arkivanov.decompose.statekeeper.Parcelize
+import com.arkivanov.essenty.parcelable.Parcelable
+import com.arkivanov.essenty.parcelable.Parcelize
 import kotlinx.serialization.Serializable
 
 @Serializable
 @Parcelize
 data class CardHistory(
-    val size: Long,
-    val actions: List<CardAction>
-): Parcelable
+    val size: Int,
+    val actions: List<CardAction>,
+) : Parcelable
